Rewrite the whole program to utilize both cores
The ESP32 microcontroller has two-core processor. This commit
implements a tally which uses both cores of the processor. The code
aims to be simple to comprehend, but at some places some tricks are
needed for safe operation in multicore environment.

The tally is almost feature-complet. Battery status and charging
mode are the only features still missing. There is also a known bug
that the device needs to be reseted between AP mode activations. I
expect that the web server doesn't free all resources when its
closed.