Add proxy for Atem
The proxy connects to Atem and serves a vMix-like interface for the
tally lights. It is easier to write a codre for a computer than a
microcontroller, so that's why it is implemented as a proxy and not
in the tally lights themselves. This also keeps the tally light
code simple and easier to audit.